After attempting to do it the "easy way" I ran into trouble with conflicts with my "flight Metrics" plugin shared function names, scripts and css. So I want back to do it the "right" way by using classes and namespaces. I started with a Wordpress plugin in boilerplate. Needed to add features to link Shortcodes. (I don't understand why that is not part of the normal boilerplate  and localize scripts.)
At this point I have the flight log working. Or at least as well as it was working with
out the boilerplate template.

The core function of this is from PGC PDP flight log system that was written in 2010
using DreamWeaver. It was converted from PHP 5.4 to 7.6. The code you see in the /Views
foler was generated from DreamWeaver. (don't blame me.)