After much frustration I finally figured out how to get a model to update that had not been reloaded from the server.
Problem: this.collection.save saved to the server and updated the model with a unique ID.
However the VIEW of the model is not updated. When I want to edit a model from the view
the form is loaded from the view. The View of the newly added model does not have the ID.
So the model can not be found to update. In this case I search for the model by yearkey.