rewrote the traditional flight log so takeoff and landing times are pure javascript/REST. Does not refresh the entire page. Also the detail page is now a javascript task and update are performed via ajax/REST. Again with no page refresh.
At this point when the detail page is used the flight that was  update is not refresh on
the main page. Was going to fix that but I need to do a commit now.

have not physically removed the old update pages. That will happen in the clean up.

Also removed some javascript libraries that were not use or redundant. (already included
in Wordpress.)